// src/app/modelos/rastreo.interface.ts
export interface RastreoI{
    id_tramite: string;
    codtramite: string;
    id_fecha_hito: string;	
    id_hito	: string;
    fecha_hito_1: string;	
    fecha_hito_2: string;	
    fecha_hito_3: string;	
    fecha_hito_4: string;    nro_hito: string;
    estado: string;
    nombre: string;
    apellido: string;
    codigo: string;
    tipo: string;
    dni: string;
    escuela: string;
    fechatramite: string;
    sexo: string;
    facultad: string;
 }